Switch to Csharpier! :D
Csharpier makes ReGitLint obsolete for most projects and I couldn't be
happier.

ReGitLint was born out of my frustration with jb cleanupcode's speed. Csharpier
is amazingly fast to the point that running it on small to medium codebases is
no problem.

Csharpier does have a few trade-offs:

* It doesn't support .editorconfig
* It's not a linter and doesn't fix code style issues
* I don't think it supports cshtml or xaml

For ReGitLint these trade-offs are okay:

* I'm okay with giving up some control over formatting for speed.
* I'm exploring combining dotnet format with csharpier as a complete solution.
* ReGitlint is all .cs

I could see ReGitLint still having a niche in the following cases:

* Projects with vb.net, razor, xaml, etc...
* Projects where ReSharper has been used for long time and reformatting
  everything is not acceptable.

Most of the projects I work on are a mix of .cs & .js. For those I'll be using
Csharpier, dotnet format, & prettier.